+++ src/main/kotlin/Telegram.kt
@@ -1,68 +1,3 @@
-import kotlinx.serialization.SerialName
-import kotlinx.serialization.Serializable
-
-@Serializable
-data class Update(
-    @SerialName("update_id")
-    val updateId: Long,
-    @SerialName("message")
-    val message: Message? = null,
-    @SerialName("callback_query")
-    val callbackQuery: CallbackQuery? = null,
-)
-
-@Serializable
-data class Response(
-    @SerialName("result")
-    val result: List<Update>,
-)
-
-@Serializable
-data class Message(
-    @SerialName("text")
-    val text: String,
-    @SerialName("chat")
-    val chat: Chat,
-)
-
-@Serializable
-data class CallbackQuery(
-    @SerialName("data")
-    val data: String,
-    @SerialName("message")
-    val message: Message? = null,
-)
-
-@Serializable
-data class Chat(
-    @SerialName("id")
-    val id: Long,
-)
-
-@Serializable
-data class SendMessageRequest(
-    @SerialName("chat_id")
-    val chatId: Long,
-    @SerialName("text")
-    val text: String,
-    @SerialName("reply_markup")
-    val replyMarkup: ReplyMarkup? = null,
-)
-
-@Serializable
-data class ReplyMarkup(
-    @SerialName("inline_keyboard")
-    val inlineKeyboard: List<List<InlineKeyboard>>,
-)
-
-@Serializable
-data class InlineKeyboard(
-    @SerialName("callback_data")
-    val callbackData: String,
-    @SerialName("text")
-    val text: String,
-)
-
 
 fun main(args: Array<String>) {
 
@@ -70,6 +5,51 @@ fun main(args: Array<String>) {
     val tbs = TelegramBotService(botToken = botToken)
     var lastUpdateId = 0L
     val trainers = HashMap<Long, LearnWordsTrainer>()
+
+    fun handleUpdate(update: Update, trainers: HashMap<Long, LearnWordsTrainer>) {
+        val message = update.message?.text
+        val chatId: Long = update.message?.chat?.id ?: update.callbackQuery?.message?.chat?.id ?: return
+        val data = update.callbackQuery?.data
+
+        val trainer = trainers.getOrPut(chatId) { LearnWordsTrainer("$chatId.txt") }
+
+        when {
+            message?.lowercase() == "/start" || data == MENU_CLICK -> {
+                tbs.sendMenu(chatId)
+            }
+
+            data == LEARN_WORDS_CLICK -> {
+                tbs.checkNextQuestionAndSend(trainer, chatId)
+            }
+
+            data?.startsWith(CALLBACK_DATA_ANSWER_PREFIX) == true -> {
+                val userAnswerIndex = data.substringAfter(CALLBACK_DATA_ANSWER_PREFIX).toInt()
+                val result = trainer.checkAnswer(userAnswerIndex)
+                if (result) tbs.sendMessage(chatId, CORRECT)
+                else tbs.sendMessage(
+                    chatId,
+                    "$NOT_CORRECT. ${trainer.question?.correctAnswer?.original} - ${trainer.question?.correctAnswer?.translate}"
+                )
+                tbs.checkNextQuestionAndSend(trainer, chatId)
+            }
+
+            data == STATS_CLICK -> {
+                val statistics = trainer.getStatistics()
+                val statisticsString =
+                    "Выучено ${statistics.learned} из ${statistics.total} слов | ${statistics.percentLearned}%"
+                tbs.sendMessage(chatId, statisticsString)
+                Thread.sleep(1000)
+                tbs.sendMenu(chatId)
+            }
+
+            data == RESET_CLICK -> {
+                trainer.resetProgress()
+                tbs.sendMessage(chatId, "Прогресс сброшен")
+                Thread.sleep(1000)
+                tbs.sendMenu(chatId)
+            }
+        }
+    }
 
     while (true) {
         Thread.sleep(2000)
@@ -79,7 +59,7 @@ fun main(args: Array<String>) {
         val response: Response = tbs.json.decodeFromString(responseString)
         if (response.result.isEmpty()) continue
         val sortedUpdates = response.result.sortedBy { it.updateId }
-        sortedUpdates.forEach { tbs.handleUpdate(it, tbs.json, trainers) }
+        sortedUpdates.forEach { handleUpdate(it, trainers) }
         lastUpdateId = sortedUpdates.last().updateId + 1
     }
 }
